Move components to proper folder
