+++ react-ui/src/common/index.js
@@ -1,4 +1,3 @@
-export { default as AppHelmet } from './AppHelmet';
 export { default as useAuthState } from './useAuthState';
 export { default as useClientWidth } from './useClientWidth';
 export { default as useRedditLogIn } from './useRedditLogIn';
@@ -6,4 +5,3 @@ export { default as useScrollState } from './useScrollState';
 export { default as useSettingsState } from './useSettingsState';
 export { default as useSwrData } from './useSwrData';
 export { default as useVotingComponentsState } from './useVotingComponentsState';
-export { default as VotingComponents } from './VotingComponents';
